Extract AnimationManager
