+++ lib/TbUiLib/CMakeLists.txt
@@ -12,6 +12,7 @@ target_sources(TbUiLib
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ActionMenu.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/AngleIndicatorRenderer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/Animation.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationManager.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/AppController.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/AppInfoPanel.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/AssembleBrushTool.cpp
@@ -258,6 +259,7 @@ target_sources(TbUiLib
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/ActionMenu.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/AngleIndicatorRenderer.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/Animation.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/AnimationManager.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/AppController.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/AppInfoPanel.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/ui/AssembleBrushTool.h
